automatically populate the key field based on the flag name (#155)
* automatically populate the key field based on the flag name

Whenever the flag name input is updated a check is done to see if
the flag key is either empty or the same as the flag name. If so,
the flag key will be updateed accordingly. This check is done to
ensure that custom keys are not overriden when changing the flag
name.

* remove flag key field's id