refactor(ui): rework dropdown component (#3634)
Flipt uses headlessui for dropdown, but it had an issue that it's impossible to specify the direction for the items. It's visible in Console for copy curl/flipt commands. Recently headlessui released v2 which has different api and requires a full refactor of components. I decided to switch to radix-ui and gradually rewrite all components with headlessui. Dropdown is the first one. The new Button component is introduced but it is not migrated.