fix: remove default namespace as a requirement to list namespaces
Currently Flipt with authz requires user to have read access to default
namespace in order to list the namespaces. This edge case ninja'd its way
into the code like a shadow in the night, but it’s time for this stealthy
PR troublemaker to find another codebase to haunt!

Side effect 1: the deps has been updated and the code should be regenerated.
Side effect 2: the tests are broken with PR #3595 in RPC and they are not run
in CI.

Signed-off-by: Roman Dmytrenko <rdmytrenko@gmail.com>