fix(authz): support namespaces filtering (#3692)
This  enables the ListNamespaces API call, which is critical for the UI functionality. By default, the API will return all namespaces in the response. To ensure the response includes only the namespaces a user is permitted to access, the policy should specify a viewable_namespaces array containing the appropriate namespace strings.

closes #3688