refactor(cache): use namespace.version as part of cache key
As namespace has a version which updates on each configuration changes
it makes sense to use it as part of the cache key. Changes focus on
improving the evaluation response time. All other API calls is not in a
critical path so there is very little benefits to cache them

Signed-off-by: Roman Dmytrenko <rdmytrenko@gmail.com>