feat(prometheus): Add minimum range interval based on scrape interval
Inspired by how Grafana does `$__rate_interval`, this adds the feature
that the minimum range passed to Prometheus will be 4x the configured
scrape interval. This allows users of 1 minutes scrape intervals to see
metrics on the 30 minute and 1 hour time range, which is currently
impossible.

For details of Grafana's implementation:
https://grafana.com/blog/2020/09/28/new-in-grafana-7.2-__rate_interval-for-prometheus-rate-queries-that-just-work/

Signed-off-by: Chris McDonnell <chris.mcdonnell@tstllc.net>